refactor(styles): name search highlight module accurately
