+++ src/shared/styles/searchHighlightStyles.ts
